layers: Dedicated cc_spirv_cooperative.cpp
