+++ BUILD.gn
@@ -142,6 +142,7 @@ vvl_sources = [
   "layers/core_checks/cc_shader_interface.cpp",
   "layers/core_checks/cc_shader_object.cpp",
   "layers/core_checks/cc_spirv.cpp",
+  "layers/core_checks/cc_spirv_cooperative.cpp",
   "layers/core_checks/cc_state_tracker.cpp",
   "layers/core_checks/cc_state_tracker.h",
   "layers/core_checks/cc_sync_vuid_maps.h",
